nlp_module/requirements.txt
fasttext==0.9.2
numpy==1.21.2
colorama==0.4.4
scipy==1.7.1
zipp==3.5.0
pandas==1.3.3
requests==2.24.0
Flask==2.0.1
virtualenv==20.8.1
flasgger==0.9.5
python-dotenv==0.19.1
compress-fasttext==0.0.7
python-Levenshtein==0.12.2

pytest==6.2.5
pytest-flask==1.2.0
